Add support for mounting FUSE filesystems in the container
Since it's not possible to completely delegate the operations to a
plugin, add support for mounting FUSE filesystems in the container
WITHOUT adding support for any specific driver or method of operation.

People wishing to support specific FUSE filesystems would need to add a
plugin that:

1. Adds a flag to specify the mount point
2. Adds a flag to specify the FUSE driver to use (along with parameters)

The reason for doing it like this is that depending on the specifics of
the FUSE filesystem that needs to be supported, more or less support
would need to be added in singularity. By adding just the necessary
support to open the /dev/fuse device and run the FUSE driver, we make
sure that we don't under- or over-specify the assumptions as to what the
driver needs or supports.

Users wishing to use FUSE filesystem in their containers would need to
make sure that the _container_ provides the necessary support to run the
driver (e.g. by shipping the driver with the container image or by
setting up bind mounts to make sure the driver is in the expected
location).

This extends the Engine configuration by adding a Plugin object which
contains entries for each of the plugins. Nothing is assumed about the
configuration. The FUSE portion *does* require that the configuration
of each plugin has a "Fuse" key pointing to an object that has _at_
least a DevFuseFd, MountPoint and Program values. Plugins that do not
support FUSE filesystems do not need to provide these keys or objects.

Signed-off-by: Marcelo E. Magallon <marcelo@sylabs.io>